Finish renaming horizonal to horizontal
...
