+++ src/ChargerAstronomyShared/Domain/Horizontal/HorizontalBody.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace ChargerAstronomyShared.Domain.Horizonal
+namespace ChargerAstronomyShared.Domain.Horizontal
 {
     public abstract class HorizontalBody
     {
