+++ src/ChargerAstronomyShared/Domain/Horizontal/HorizontalPlanet.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace ChargerAstronomyShared.Domain.Horizonal
+namespace ChargerAstronomyShared.Domain.Horizontal
 {
     public class HorizontalPlanet : HorizontalBody
     {
